Fix active tab during different states
Praise SchizoDuckie.

Context:
Library tab is only active when in the library which is the 'favorites'
state.
Add a Show is active when adding a show from trending or search. Both
trending and search pages belong under the 'favorites.add' so if we're
on favorites.add.trending or favorites.add.search , the tab will remain
active.
A problem is when you hover over a serie, the active state changes to
the 'trakt-serie' which is the sidepanel that shows serie details so we
also have to check if we're in the 'trakt-serie' state. Seeing as this
state can only be active while in add mode it's fine.
And filtering uses 'favorites.search' although i'm unsure if I'll use
the this method for filtering.